Configure Git to auto sign every commit
I was using an alias to sign my commits,
but I just found out about commit.gpgSign.
So I removed the alias and set git to sign
everything (including rebases!).

PHPStorm seems to not be able to handle this
very well, so I needed to add the no-tty
option to the gpg.conf file to avoid errors
when commiting from the IDE.