Improve error handling in currency-converter
Sometimes, the API will not return a valid JSON file,
so jq was updated to check the the currency is included
in the response and, if not, it will print an empty string
instead of null (which is what it prints when it can find
something matching the query). An empty string will not
break the printf statement, but the null will.

Also, errors from the curl call are now redirected to
/dev/null